+++ src/onegov.election_day/onegov/election_day/tests/test_models.py
@@ -230,7 +230,8 @@ def test_archive(session):
     assert archive.for_date(2015).date == 2015
 
     assert archive.get_years() == []
-    assert archive.latest() is None
+    assert archive.latest() == []
+    assert archive.for_date(2015).by_date() == []
 
     for year in (2009, 2011, 2014, 2016):
         session.add(
@@ -254,20 +255,23 @@ def test_archive(session):
 
     assert archive.get_years() == [2016, 2015, 2014, 2011, 2009, 2007]
 
-    assert archive.latest() == archive.for_date(2016).by_date()
-    assert archive.latest() == archive.for_date('2016').by_date()
-    assert archive.latest() == archive.for_date('2016-01-01').by_date()
+    for date_ in (2016, '2016', '2016-01-01'):
+        assert archive.latest() == archive.for_date(date_).by_date()
 
-    assert archive.for_date('2016-02-02').by_date() is None
+    assert archive.for_date('2016-02-02').by_date() == []
 
     for year in (2009, 2011, 2014, 2016):
-        assert (
-            ('election', 'federation', date(year, 1, 1)),
-            [session.query(Election).filter_by(date=date(year, 1, 1)).one()]
-        ) in archive.for_date(year).by_date()
+        item = session.query(Election).filter_by(date=date(year, 1, 1)).one()
+        items = archive.for_date(year).by_date()
+        assert item in items
+
+        groups = archive.group_items(items)
+        assert groups[date(year, 1, 1)]['federation']['election'] == [item]
 
     for year in (2007, 2011, 2015, 2016):
-        assert (
-            ('vote', 'federation', date(year, 1, 1)),
-            [session.query(Vote).filter_by(date=date(year, 1, 1)).one()]
-        ) in archive.for_date(year).by_date()
+        item = session.query(Vote).filter_by(date=date(year, 1, 1)).one()
+        items = archive.for_date(year).by_date()
+        assert item in items
+
+        groups = archive.group_items(items)
+        assert groups[date(year, 1, 1)]['federation']['vote'] == [item]